concurrency has reached a beta status. Tasks-syncing works realtime (read-mode) - conflict management for concurrently edited tasks are implemented (data, ui). NEXT STATIONS: realtime sync in write-mode; ui for managing concurrently deleted tasks; removing own conflicted-management data storage (dataremote) and using build-in capabilities (conflicted_rev)

+++ evently/projects/loadProjects/selectors/a.start-sync/click.js
@@ -1,22 +1,22 @@
 function () {
 	
-	console.log("start sync");
-
-	$("#sync_state").text("Start syncing ...");
+	$("#messagebox").text("Start syncing ...");
 	
-	var widget = $$("#project");
+	var widget = $$("#sync_state");
 	var db = widget.app.db;
 	var dinos = widget.graph.dirtyNodes();
 	var lastd = dinos.last();
 	var localgraph = JSON.parse(localStorage.graph);
 
-	$("#sync_state").text("looking for local changes ...");
+	$("#messagebox").text("looking for local changes ...");
 	dinos.each(function (doc) {
 		
 		var docjson = doc.toJSON();
 
 		if (doc._deleted == true) {
-			db.removeDoc(docjson);
+			if(doc._rev !== undefined) {
+				db.removeDoc(docjson);
+			}
 			delete localgraph[doc._id];
 		} else {
 			db.saveDoc(docjson, {
@@ -30,13 +30,13 @@ function () {
 
 		if(doc._id === lastd._id) {
 			localStorage.graph = JSON.stringify(localgraph);
-			$("#sync_state").text("synced!");
+			$("#messagebox").text("synced!");
 			
 
 			//TODO this event has to be called after the last doc update (deleted doc dont matter)
 			//the events renders the view with the synced data and must chained after that
-			var projectid = $("a.active").attr("project");
-			$(this).trigger("loadProject", [{"projectid": projectid}]);
+			//var projectid = $("a.active").attr("project");
+			//$(this).trigger("loadProject", [{"projectid": projectid}]);
 		}
 
 	});
